Move `Domainic::Type::Constraint::ParameterSet`
This moves `Domainic::Type::Constraint::ParameterSet` to
`Domainic::Type::Constraint::Specification::ParameterSet`
